+++ packages/orq-rc/src/models/components/conversationresponse.ts
@@ -107,7 +107,7 @@ export const ConversationResponse$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01khrv27x6a3atnmz5vjyk8aa9"),
+  _id: z.string().default("conv_01kht0v9xenpgat2x942mq37zj"),
   kind: ConversationResponseKind$inboundSchema,
   displayName: z.string(),
   createdAt: z.number(),