+++ packages/orq-rc/src/models/components/conversationwithmessagesresponse.ts
@@ -122,7 +122,7 @@ export const ConversationWithMessagesResponse$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01khrv27xcqdypncyqds80bgeh"),
+  _id: z.string().default("conv_01kht0v9xmx06pxdhhphr5ggdg"),
   kind: ConversationWithMessagesResponseKind$inboundSchema,
   displayName: z.string(),
   createdAt: z.number(),