+++ packages/orq-rc/src/models/components/partdoneevent.ts
@@ -76,7 +76,7 @@ export const PartReasoningPart$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("reasoning_01khrv27jb0kss0em42sb8s42h"),
+  _id: z.string().default("reasoning_01kht0v9hp1sfaf6smfjnxwkjt"),
   metadata: z.record(z.any()).optional(),
   kind: PartKind$inboundSchema,
   reasoning: z.string(),