+++ packages/orq-rc/src/models/operations/createconversationresponse.ts
@@ -280,7 +280,7 @@ export const PartReasoningPart$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("reasoning_01khrv27y0d12h8w9fgz34fy8h"),
+  _id: z.string().default("reasoning_01kht0v9yjg4gr56hhyh8c5wwv"),
   metadata: z.record(z.any()).optional(),
   kind: PartKind$inboundSchema,
   reasoning: z.string(),
@@ -366,7 +366,7 @@ export const ReasoningPart$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("reasoning_01khrv27xxzc8w52naehmt1572"),
+  _id: z.string().default("reasoning_01kht0v9yedez7b5hen0vr6g0c"),
   metadata: z.record(z.any()).optional(),
   kind: DeltaKind$inboundSchema,
   reasoning: z.string(),