+++ packages/orq-rc/src/models/operations/generateconversationname.ts
@@ -180,7 +180,7 @@ export const GenerateConversationNameResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01khrv27xanex1yzpyky6m43wf"),
+  _id: z.string().default("conv_01kht0v9xjwwnfb4zrad598k1q"),
   kind: GenerateConversationNameKind$inboundSchema,
   displayName: z.string(),
   createdAt: z.number(),