+++ packages/orq-rc/src/models/operations/listdatasources.ts
@@ -176,7 +176,7 @@ export const ListDatasourcesData$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KHRV28DQD964Y9B705XDPN5N"),
+  _id: z.string().default("01KHT0VAEDDPEBXNN3C5NT2HQT"),
   display_name: z.string(),
   description: z.string().optional(),
   status: ListDatasourcesStatus$inboundSchema,