+++ packages/orq-rc/src/models/operations/retrievedatasource.ts
@@ -118,7 +118,7 @@ export const RetrieveDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KHRV28DS58YFFVX5JDNY546M"),
+  _id: z.string().default("01KHT0VAEFMWQJGMKFKVXHPS08"),
   display_name: z.string(),
   description: z.string().optional(),
   status: RetrieveDatasourceStatus$inboundSchema,