+++ packages/orq-rc/src/models/operations/updateconversation.ts
@@ -226,7 +226,7 @@ export const UpdateConversationResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01khrv27xjeb32gy9vvznq7h1r"),
+  _id: z.string().default("conv_01kht0v9y4wssgp2e8jq5rve0e"),
   kind: UpdateConversationKind$inboundSchema,
   displayName: z.string(),
   createdAt: z.number(),